requirements.txt
albumentations==1.0.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.1.1
cycler==0.12.1
einops==0.8.0
fonttools==4.54.1
idna==3.10
imageio==2.35.1
importlib_resources==6.4.5
kiwisolver==1.4.7
lazy_loader==0.4
matplotlib==3.7.5
networkx==3.1
numpy==1.24.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==10.4.0
pip==24.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytorchcv==0.0.65
PyWavelets==1.4.1
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.21.0
scipy==1.10.1
setuptools==75.1.0
six==1.16.0
tifffile==2023.7.10
typing_extensions==4.12.2
urllib3==2.2.3
wheel==0.44.0
zipp==3.20.2
